requirements.txt
aiogram==3.4.1
psycopg[binary]==3.2.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-docx==1.1.2
pdfplumber==0.11.4
requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1
